05_Agent_with_Chainlit_Ui/pyproject.toml
[project]
name = "agent-with-chainlit-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.4.400",
    "openai-agents==0.0.7",
]
